Dissociate logging from events
